endOfLesson#5RepeatPart#1

+++ src/services/user.service.ts
@@ -1,21 +1,49 @@
-import { IUser, IUserDTO } from "../interfaces/user.interface";
+import { StatusCodesEnum } from "../enums/status-codes.enum";
+import { ApiError } from "../errors/api.error";
+import {
+    IUser,
+    IUserCreateDTO,
+    IUserUpdateDTO,
+} from "../interfaces/user.interface";
 import { userRepository } from "../repositories/user.repository";
 
 class UserService {
     public getAll(): Promise<IUser[]> {
         return userRepository.getAll();
     }
-    public create(user: IUserDTO): Promise<IUser> {
+    public create(user: IUserCreateDTO): Promise<IUser> {
         return userRepository.create(user);
     }
-    public getById(id: string): Promise<IUser> {
-        return userRepository.getById(id);
+    public async getById(id: string): Promise<IUser> {
+        const user = await userRepository.getById(id);
+
+        if (!user) {
+            throw new ApiError("user not found", StatusCodesEnum.NOT_FOUND);
+        }
+        return user;
     }
-    public updateById(id: string, user: IUserDTO): Promise<IUser> {
-        return userRepository.updateById(id, user);
+    public async updateById(id: string, user: IUserUpdateDTO): Promise<IUser> {
+        const data = await userRepository.getById(id);
+        if (!data) {
+            throw new ApiError("user not found", StatusCodesEnum.NOT_FOUND);
+        }
+        return await userRepository.updateById(id, user);
     }
-    public deleteById(id: string): Promise<IUser> {
-        return userRepository.deleteById(id);
+    public async deleteById(id: string): Promise<void> {
+        const data = await userRepository.getById(id);
+        if (!data) {
+            throw new ApiError("user not found", StatusCodesEnum.NOT_FOUND);
+        }
+        await userRepository.deleteById(id);
+    }
+    public async isEmailUnique(email: string): Promise<void> {
+        const newUser = await userRepository.getByEmail(email);
+        if (newUser) {
+            throw new ApiError(
+                "Such user already exists",
+                StatusCodesEnum.BED_REQUEST,
+            );
+        }
     }
 }
 export const userService = new UserService();
